[BACKLOG-18737]-Resolve issue with BundleNotFoundException org.pentaho.big.data.kettle.plugins.formats during spark run, fixed error on first load, the error was connected with having 2 packages with one name both in formats and formats-meta packages, for api and impl they should be different, part of fix in scope of pentaho-ee
