+++ kettle-plugins/formats/src/main/java/org/pentaho/big/data/kettle/plugins/formats/impl/avro/input/AvroInputData.java
@@ -20,7 +20,7 @@
  *
  ******************************************************************************/
 
-package org.pentaho.big.data.kettle.plugins.formats.avro.input;
+package org.pentaho.big.data.kettle.plugins.formats.impl.avro.input;
 
 import java.util.Iterator;
 import java.util.List;
