+++ kettle-plugins/formats/src/main/java/org/pentaho/big/data/kettle/plugins/formats/impl/avro/input/AvroInputDialog.java
@@ -19,9 +19,7 @@
  * limitations under the License.
  *
  ******************************************************************************/
-
-package org.pentaho.big.data.kettle.plugins.formats.parquet.input;
-
+package org.pentaho.big.data.kettle.plugins.formats.impl.avro.input;
 
 import org.eclipse.swt.SWT;
 import org.eclipse.swt.widgets.Button;
@@ -32,7 +30,8 @@ import org.eclipse.swt.widgets.Listener;
 import org.eclipse.swt.widgets.Shell;
 import org.eclipse.swt.widgets.TableItem;
 import org.pentaho.big.data.kettle.plugins.formats.FormatInputField;
-import org.pentaho.big.data.kettle.plugins.formats.parquet.BaseParquetStepDialog;
+import org.pentaho.big.data.kettle.plugins.formats.impl.avro.BaseAvroStepDialog;
+import org.pentaho.big.data.kettle.plugins.formats.avro.input.AvroInputMetaBase;
 import org.pentaho.di.core.Const;
 import org.pentaho.di.core.row.value.ValueMetaFactory;
 import org.pentaho.di.i18n.BaseMessages;
@@ -40,9 +39,8 @@ import org.pentaho.di.trans.TransMeta;
 import org.pentaho.di.ui.core.widget.ColumnInfo;
 import org.pentaho.di.ui.core.widget.ColumnsResizer;
 import org.pentaho.di.ui.core.widget.TableView;
-import org.pentaho.hadoop.shim.api.format.SchemaDescription;
 
-public class ParquetInputDialog extends BaseParquetStepDialog<ParquetInputMeta> {
+public class AvroInputDialog extends BaseAvroStepDialog<AvroInputMetaBase> {
 
   private static final int DIALOG_WIDTH = 526;
 
@@ -56,36 +54,32 @@ public class ParquetInputDialog extends BaseParquetStepDialog<ParquetInputMeta> 
 
   private TableView wInputFields;
 
-  public ParquetInputDialog( Shell parent, Object in, TransMeta transMeta, String sname ) {
-    super( parent, (ParquetInputMeta) in, transMeta, sname );
+  public AvroInputDialog( Shell parent, Object in, TransMeta transMeta, String sname ) {
+    super( parent, (AvroInputMetaBase) in, transMeta, sname );
   }
 
   @Override
   protected Control createAfterFile( Composite shell ) {
 
     Button wGetFields = new Button( shell, SWT.PUSH );
-    wGetFields.setText( BaseMessages.getString( PKG, "ParquetInputDialog.Fields.Get" ) );
+    wGetFields.setText( BaseMessages.getString( PKG, "AvroInputDialog.Fields.Get" ) );
     wGetFields.addListener( SWT.Selection, event -> {
-      try {
-        setFields( ParquetInput.retrieveSchema( meta.namedClusterServiceLocator, meta.getNamedCluster(), wPath.getText()
-            .trim() ) );
-      } catch ( Exception ex ) {
-        throw new RuntimeException( ex );
-      }
+      // TODO
+      throw new UnsupportedOperationException();
     } );
     props.setLook( wGetFields );
     new FD( wGetFields ).bottom( 100, 0 ).right( 100, 0 ).apply();
 
     Label wlFields = new Label( shell, SWT.RIGHT );
-    wlFields.setText( BaseMessages.getString( PKG, "ParquetInputDialog.Fields.Label" ) );
+    wlFields.setText( BaseMessages.getString( PKG, "AvroInputDialog.Fields.Label" ) );
     props.setLook( wlFields );
     new FD( wlFields ).left( 0, 0 ).top( 0, FIELDS_SEP ).apply();
     ColumnInfo[] parameterColumns = new ColumnInfo[] {
-      new ColumnInfo( BaseMessages.getString( PKG, "ParquetInputDialog.Fields.column.AvroPath" ),
+      new ColumnInfo( BaseMessages.getString( PKG, "AvroInputDialog.Fields.column.AvroPath" ),
           ColumnInfo.COLUMN_TYPE_TEXT, false, false ),
-      new ColumnInfo( BaseMessages.getString( PKG, "ParquetInputDialog.Fields.column.Name" ),
+      new ColumnInfo( BaseMessages.getString( PKG, "AvroInputDialog.Fields.column.Name" ),
           ColumnInfo.COLUMN_TYPE_TEXT, false, false ),
-      new ColumnInfo( BaseMessages.getString( PKG, "ParquetInputDialog.Fields.column.Type" ),
+      new ColumnInfo( BaseMessages.getString( PKG, "AvroInputDialog.Fields.column.Type" ),
           ColumnInfo.COLUMN_TYPE_CCOMBO, ValueMetaFactory.getValueMetaNames() ) };
     wInputFields =
         new TableView( transMeta, shell, SWT.FULL_SELECTION | SWT.SINGLE | SWT.BORDER | SWT.NO_SCROLL | SWT.V_SCROLL,
@@ -110,7 +104,7 @@ public class ParquetInputDialog extends BaseParquetStepDialog<ParquetInputMeta> 
    * Read the data from the meta object and show it in this dialog.
    */
   @Override
-  protected void getData( ParquetInputMeta meta ) {
+  protected void getData( AvroInputMetaBase meta ) {
     if ( meta.inputFiles.fileName.length > 0 ) {
       wPath.setText( meta.inputFiles.fileName[0] );
     }
@@ -134,26 +128,15 @@ public class ParquetInputDialog extends BaseParquetStepDialog<ParquetInputMeta> 
     }
   }
 
-  protected void setFields( SchemaDescription schema ) {
-    wInputFields.table.removeAll();
-    for ( SchemaDescription.Field f : schema ) {
-      TableItem item = new TableItem( wInputFields.table, SWT.NONE );
-
-      item.setText( AVRO_PATH_COLUMN_INDEX, f.formatFieldName );
-      item.setText( FIELD_NAME_COLUMN_INDEX, f.formatFieldName );
-      item.setText( FIELD_TYPE_COLUMN_INDEX, ValueMetaFactory.getValueMetaName( f.pentahoValueMetaType ) );
-    }
-  }
-
   /**
    * Fill meta object from UI options.
    */
   @Override
-  protected void getInfo( ParquetInputMeta meta, boolean preview ) {
+  protected void getInfo( AvroInputMetaBase meta, boolean preview ) {
     String filePath = wPath.getText();
     if ( filePath != null && !filePath.isEmpty() ) {
       meta.allocateFiles( 1 );
-      meta.inputFiles.fileName[0] = wPath.getText().trim();
+      meta.inputFiles.fileName[0] = wPath.getText();
     }
     int nrFields = wInputFields.nrNonEmpty();
     meta.inputFields = new FormatInputField[nrFields];
@@ -179,7 +162,7 @@ public class ParquetInputDialog extends BaseParquetStepDialog<ParquetInputMeta> 
 
   @Override
   protected String getStepTitle() {
-    return BaseMessages.getString( PKG, "ParquetInputDialog.Shell.Title" );
+    return BaseMessages.getString( PKG, "AvroInputDialog.Shell.Title" );
   }
 
   @Override
