+++ kettle-plugins/formats/src/main/resources/org/pentaho/big/data/kettle/plugins/formats/impl/avro/input/messages/messages_en_US.properties
