+++ kettle-plugins/formats/src/main/resources/org/pentaho/big/data/kettle/plugins/formats/impl/avro/output/messages/messages_en_US.properties
